refactor into a static maintenance page function called many times
